rename lib dir to libs
