+++ modules/api-explorer-client/src/libs/phenylClient.ts
@@ -1,0 +1,15 @@
+import PhenylHttpClient from "@phenyl/http-client";
+
+const { phenylApiUrlBase } = window.phenylApiExplorerClientGlobals;
+
+let client: any | null = null;
+
+// Singleton PhenylHttpClient
+export default function getPhenylHttpClient() {
+  if (!client) {
+    client = new PhenylHttpClient({
+      url: phenylApiUrlBase || window.location.origin
+    });
+  }
+  return client;
+}
