+++ modules/api-explorer-client/src/modules/user.ts
@@ -1,6 +1,6 @@
-/* global PhenylFunctionalGroupSkeleton */
 import { ThunkDispatch } from "redux-thunk";
-import getPhenylHttpClient from "../lib/phenylClient";
+
+import getPhenylHttpClient from "../libs/phenylClient";
 
 const { PhenylFunctionalGroupSkeleton } = window.phenylApiExplorerClientGlobals;
 
@@ -101,7 +101,6 @@ export const login = (
   // @TODO: define those any typs
 ) => async (dispatch: ThunkDispatch<any, any, any>) => {
   const client = getPhenylHttpClient();
-  console.log(client);
 
   try {
     dispatch(loginRequest());
